Add parent_content_id to Dimensions::Editions
We need the content_id of the parent manual that a manual_section
belongs to in order to generate the link to the Manuals Publisher.

Rather that store this unnecessarily in the database to handle manuals
and manuals_section as an edge case, sending that as part of metadata
is slightly better where most things would have this set to nil.

In future if we need to store the parent_content_id then we can refactor
it, or indeed implement this better.