Remove /document_types in favor of /api/v1/document_types
We no longer need /document_types because it is currently covered with
/api/v1/document_types. In the existing implementation we had two ways
to return the document types:

1. Using a finder
2. Use a method within the DocumentType

I have decided to continue with the existing approach (Finders/Queries)
so we have the code is more consistent with the existing ways to retrieve
information.

I have moved tests and code into the finder, and removed previous
implementation.