Remove /organisations in favor of /api/v1/organisations
We no longer need /organisations because it is currently covered with
/api/v1/organisations. In the existing implementation we had two ways
to return the document types:

1. Using a finder
2. Use a method within the Orgnisation

I have decided to continue with the existing approach (Finders/Queries)
so we have the code is more consistent with the existing ways to retrieve
information.

I have moved tests and code into the finder, and removed previous
implementation.