Make configurable the Icinga alarm validation
We have different schedules for jenkins job that runs the ETL master
process, for this reason we need to run the validations at different
times. On top of this, we also want to rollout the validations through
the environments in multiple iterations, so I have also made configurable
the notifications to 2nd line via the healthcheck endpoint.

Related PRs:

1. https://github.com/alphagov/govuk-puppet/pull/8619
2. https://github.com/alphagov/govuk-puppet/pull/8616