Add migration to update reading time on latest items
Previous commit added reading time field and calculation. This commit
updates the Facts::Edition reading_time field for latest
Dimensions::Edition content items. We don't need to update items in the
past as we don't calculate time series data from this metric. This also
has the benefit of speeding up the migration as it is a lot of data to
run through otherwise.