Update total_results calculation
Users are seeing a discrepancy in the number provided by
total_results and the actual total returned.

This scenario occurs when a content item has been newly published, but
the ETL process has not yet run. The result set will exclude the new item
until the ETL process has run by the next day, but the total_result
still counts that Edition.