Use `warehouse_item_id` to guarantee order in /content endpoint
We can’t rely on attributes that contain duplicates to sort the data. 
The duplication in those values can make our pagination to return the 
same rows multiple times.

This is the case for content items with no views. For example, if we 
List all the HTML publications, we have more than 8K content items with 
no views so pagination through them won’t be consistent.