Chomp off long page_paths from GA reports
During the ETL process, page_paths that have very long query params
(suspicious looking query stuffing) will cause the import to fail as the
index can't store a string value so long. This pragmatic fix will allow
the ETL to complete, as we do not treat URLs with query params as
separate instances of page visits as Google Analytics does so we can
mitigate against this problem.