Enable null values for satisfaction score
Null values are a valid representation of satisfaction scores, as it
denotes that we did not get any responses for the yes/no survey. This
also helps to fix the representation of data in the graphs, which until
now have been misleading - setting 0 as a satisfacion score (the
default value) for periods even before the rollout of the yes/no
survey is misleading.