Ignore pages with long query strings from GA
Previously we were using slice. Slice uses the character at that
position of the string and caused multiple pages to have the same
page_path which violated the GA Events table primary key
constraint. This is a quick fix, based on the assumption long base
paths are due to long query strings and the fact we don't use data
about paths with query strings. We should look for a solution to
deal with query strings.