Add type casts to aggregation views
This adds integer type casts to the aggregations views as there
is no way to specify column types for materialized view.
We currently cast the type in the Content finder. Moving the type
cast to postgres, removes complexity from the finder and helps
support batch quering need for CSV requests. The columns were
previously being cast as Numeric due to the Postgres SUM function
returning Numeric when the arguments are Bigints.