Add healthcheck for SearchLastMonth ETL
We generate materialised views for searching acrosss the content. A
refresh of the views is triggered when the latest Dimensions::Edition
items are generated. This tests that the ETL process for this occurred.

So far we have added a healthcheck for last_thirty_days and last_month.
Both of these are near identical, so the next commit will refactor.