Add updated_at to all other search aggregation views
An earlier commit added this column to the last 30 days view, with a
subsequent test to ensure that the healthcheck works. This commit adds
the column to the the other search views.

We need to check the freshness of the materialized view as part of
healthchecks. The materialized views are supposed to be refrehsed after
the load stage of the ETL process. We need to check that these ran ok.