Fix the Healthchecks::SearchAggregations test to be more consistent
Previously, it worked for the first half of the month, as the metric
created 15.days.ago was in the previous month. However, this didn't
work in the later half of the month.

Switch to placing the created metrics always in the last month, so
that the test passes throughout the month.

Also call recalculate_aggregations! in the second context block, as
I'm not sure this happens automatically, and it might be necessary.