Change message handlers to find latest edition
This updates the single and multipart message handlers to use the
find latest edition method to find the most recent edition.
Previously, the single message handler looked for the `live`
edition. As we are redefining `live` to only be true for content
live on GOV.UK not all content items will have a `live` edition.
We also making the logic consistent with multipart hander, which
previously looked for edition which was `live` for the base path.