Change trap and time_and_trap to return success as a boolean
These functions are intended to allow the ETL process to continue, if
a part of it fails, but it's important to know if any part has failed,
so that the Rake task aborts, to make Jenkins/Icinga aware that there
has been an issue.

Therefore, return a boolean indicating success.