Stop catching exceptions in Etl::Feedex::Processor.extract_events
Errors will therefore be caught in process, which will indicate this
to the caller by the return value.

This does slightly change the behaviour here, as previously, the
`load_metrics` step would take place, even if there were failures
importing the events. It's simpler to just abort the entire process,
so lets do that for now, as I'm not sure there's a reason to continue
with partial data.