Check for failures, and abort if detected for ETL Rake tasks
Previously, due to the error capturing in the ETL process, failures
would be logged, but the Rake task would exit indicating that nothing
went wrong.

The exit code is displayed in the status of the Jenkins job, and
therefore also used in the Icinga monitoring.

This change should ensure that the Jenkins job looks like it's failed
if the process encounteres an issue, and that the passive Icinga
monitoring also shows the failure.