Fix find series to use warehouse item id
The intention for the API was to display metrics for a content
item over time and provide support content items that may change
their base path. Previously the find series function collected
metrics associated with a base path. This commit changes this to
warehouse item id as previous editions may have used a different
base path.