Trap errors in our Sidekiq jobs
The default behaviour for the Sidekiq jobs is
the write the whole exception including a full stack
trace into the logs.

This is causing problems when we have an error in our code
because the huge log files fill the hard disk, causing problems
for the other apps on the `backend` machines.

This commit traps the errors in our own code and:

* Writes only the message from the exception to the logs
* Sends the entire exception to sentry, where we have set up a Slack
  notification for recurring errors.