Fix the last three months aggregations spec
It was broken today and yesterday, as Ruby 3.months.ago was the 28th
of Feb, and this didn't get included in the materialised view.

At least in Content Data, the date displayed for this is from the 1st
of March, so add a day to the date for the metric, so that it lies
within the last three months.