Store organisations information in dimensions_editions
This is a precursor to adding this same information to the
materialized views, which should enable augmenting the existing
organisation filtering to take this information in to account.

Currently, there are issues when you filter for some organisations, as
some of their content is published by GDS, and thus doesn't
appear. Including the additional organisations from the organisations
link type should improve the results returned.