Add child_sort_order to dimensions_editions
Adds a column to cache the order of child editions within
a parent edition.

This is because we receive parents/children separately from
the publishing api.

The only time we know the order of the sections is when we are processing
the parent (from links -> sections).

We need to store this information so we can apply the correct sort order.