Mark 4 coming soon editions as not live
These editions shouldn't be live, as they should have been replaced by
some actual content. However, due to the special behaviour in the
Publishing API for replacing coming_soon items, the Content Data API
wasn't informed.

The plan to rectify this is to manually mark the affected items as not
live (this migration), and then manually have the Publishing API send
the Content Data API the state of the affected pages.