Allow for known_schemas, which don't exist
Currently, this test is run on govuk-content-schemas changes. Also,
the content-data-api tests are run with the govuk-content-schemas that
are deployed to Production.

Without this change, you can't add new schemas, as the Content Data
API tests will fail with the new schema in the list of known schemas,
and with it omitted.

To allow adding new schemas without bypassing the tests, just check
that all existing schemas are known, and ignore any which are in the
known_schemas list, but that don't exist yet.