Update where the government details can be found
Work is underway to link content to it's associated government, which
removes the need to have information about the government in the
details of the content.

This will allow the Publishing API to take care of updating the
content items when the government becomes historical.

Once all relevant content is linked to the right government, this
change can be merged to ensure the Content Data API has access to up
to date information.