Add support for aggregations by specific month and year
A request can now include a date_range param such as "november-2019"
to return data for that month. If the date range is such a format it
will return the `aggregations_monthly_metrics` table and MonthlyMetric
model rather than the preset data ranges.

An additional filter is added to then join this table with the
`dimensions_editions` and `facts_editions` to access the relevant data.