Add missing conditional for 'message' value
Other health checks in this folder have this conditional,
so I assume this one got missed.

Without it, we're getting this in the alert:
'daily_metrics - ok - ETL :: no daily metrics for yesterday'

...which is a confusing message, as it says it's OK but also
that there are no daily metrics for yesterday. In reality,
there ARE daily metrics and we are just incorrectly returning
a message.

Trello: https://trello.com/c/2wiNWp19/2065-investigation-content-data-api-app-healthcheck-not-ok-timebox-2-day