Call GovukError.notify with IDs of old and new editions
This diagnostic code was originally added in the wrong place:
https://github.com/alphagov/content-data-api/pull/1522/files

Currently it is hard to know exactly which Dimensions::Edition
is being updated - no edition ID exists in Sentry:
https://sentry.io/organizations/govuk/issues/1517959668/?environment=production&project=1461890&query=is%3Aunresolved

This commit attempts to add more diagnostic information, so
that we know exactly what payload has caused the error.
It removes the code from the place it was incorrectly added to,
and edits the test to reflect reality. There is no corresponding
test for base_handler.rb, and it didn't seem worthwhile to add
one given its value and the complexity of testing a private
method.

I've checked and the RecordNotUnique error does not get filtered
out upstream, so we are safe to pass this to GovukError.notify:
https://github.com/alphagov/govuk_app_config/blob/master/lib/govuk_app_config/configure.rb

Trello card: https://trello.com/c/2wiNWp19/2065-investigation-content-data-api-app-healthcheck-not-ok-timebox-2-day