Remove 'before' blocks to make tests clearer
'before' blocks should be used for setup code. They should
not be used to perform actions on the subject that we then
assert against in the tests themselves.
I think someone wanted to DRY the tests up but the result
is harder to comprehend, and harder to add to.

In the next commit I'll be adding a test which relies on
mocking the behaviour that would otherwise have been
swallowed up in the 'before' block.