WIP: write a failing test
This attempts to recreate the issue we're seeing on Production.
We have a HtmlAttachment that was Unpublished (successfully)
and then seemingly republished, possibly due to the
RepresentDownstream events that occurred afterwards.

Or, possibly (more likely) caused by the PutContent/PatchLinkSet/
Publish events of another content item that has the same
base_path. However, we already have a test that appears to check
against that scenario:
https://github.com/alphagov/content-data-api/blob/28aadabd97b24f9cfcb4b16df1c8cfa1711d3597/spec/integration/streams/single/publishing_editions_spec.rb#L248-L288

Trello: https://trello.com/c/2wiNWp19/2065-investigation-content-data-api-app-healthcheck-not-ok-timebox-2-day#comment-5f43cc60fb745674b10022a4