Account for leap year in 'day of year' validation
The build is failing today (31st December 2020) because 2020 is a
leap year, therefore this year there are 366 days in the year
rather than 365.

I did try conditionally validating the range `1..365` or `1..366`
depending on the outcome of `Date.leap?(year)`, but because of
_when_ the validation happens, `year` is often `nil`, so is not
a reliable attribute to test against. I don't foresee any major
issues by simply allowing the number 366, even if the given year
is not a leap year.