Add govuk_test
This is a library that we should be using in all our applications to
provide a consistent test environment. In this particular case, it
brings in a version of Brakeman that won't be automatically updated to
the latest version and thus avoids problems of PRs failing due to an
external factor.

We should eventually upgrade to a more recent version of Brakeman,
however this should be done in coordination with the govuk_test Gem so
it's something we do across all our apps at the same time.

See https://github.com/alphagov/govuk_test/pull/28 for more background
on why we're doing this.