Run `rails app:update` to get 6.1 configuration
Carefully worked through the changes and committed only those that
make sense for Content Data API.

One thing worth noting is that the `config.file_watcher` value
in `development.rb` has changed by default. The reason it was
overwritten in the first place was down to the VM (see
9d206989443a1ea968d8cf9ad41a0f258345fb47), which is no longer
relevant, so I've allowed the Rails default to take hold here.