Remove pointless test
This test checked that `GovukSchemas::Schema.schema_names` included all
the known schema names.

When adding a new schema, this meant we had to add it to the list in
this test without making any changes to the application code.

This repo is not an appropriate place for that test, as it solely checks
the behaviour of a gem which has its own test for the method [1].

Removing the test will mean `content-data-api` won't need to be updated
each time we add a new schema type.

1: https://github.com/alphagov/govuk_schemas/blob/4fef0aa95be24c3534a00b622a59ea7f1e8175fa/spec/lib/schema_spec.rb#L30-L36