Stop collecting unused GA metrics on pages
This commit removes reference to, and the collecting of:
- bounces
- entrances
- exits
- page_time / average page_time

These metrics were introduced during the development of Content Data
but were not fully implemented as the team building it was disbanded
whilst it was in it's Beta stage.

There are no references to them in content-data-admin and are not
available to the user through the UI.

These metrics will no longer be collected going forward as we migrate
away from UA to GA4/BigQuery. By default these metrics are set to 0
when entries are being made to the database.

Further down the line after the migration work is done, we can look at
removing these fields from the database once we are confident we don’t
need them.