Set up BigQuery connection
The GA4 API does not have direct mappings available for the
metrics we currently collect from UA. Our workaround is to use
BigQuery which provides more flexibility and futher processing
of GA4's data that we can query.

This commit tests the initial connection, and the class will be
expanded in future commits as we migrate each service from UA
to GA4:

- Views & Navigation
- Internal Search
- User Feedback

BigQuery requires the following credentials for authentication:
- `project_id`
- `private_key` (stored in the credentials JSON)
- `client_email` (stored in the credentials JSON)

The credentials will be accepted in the form of a file or as JSON.
We can store them in environment variables which are then converted
into the correct format. For more information see https://cloud.google.com/ruby/docs/reference/google-cloud-bigquery/latest/AUTHENTICATION

This is configured locally in local_env.yml and stored in
AWS Secrets Manager.