Update Views & Navigation service to pull metrics from GA4/BigQuery
The metrics required will be pre-stored in a table in BigQuery. It
will contain data from when GA4 was first made available (since 2022),
and will be configured to pull in data from the past 24hrs daily at
~3am where there aren't many users on the site. This may need to be
tweaked later to ensure the data is available before the scheduled
cron tasks are run in each environment (https://github.com/alphagov/govuk-helm-charts/blob/main/charts/app-config/values-production.yaml#L499-L580)

Instead of fetching data from UA we now use the BigQuery API which
allows us to query the table for the specific metrics we need. It
already returns the information we need in the correct format and
reduces the amount of work required post query.