Update ETL processor to get daily metrics from 2 days ago
We are migrating to using Google's GA4/BigQuery service as they are
switching off UA, which is what Content Data currently uses to collect
metrics on GOV.UK pages.

The GA4 product currently has a 48 hours lag to finalise the data and
be available to access in BigQuery, as opposed to UA which is a day.