Replace deprecated google-api-client gem
This fixes the following deprecation warning and error by
replacing the gem with the specific service used in the app:

```
*******************************************************************************
The google-api-client gem is deprecated and will likely not be updated further.

Instead, please install the gem corresponding to the specific service to use.
For example, to use the Google Drive V3 client, install google-apis-drive_v3.
For more information, see the FAQ in the OVERVIEW.md file or the YARD docs.
*******************************************************************************
```

```
Failure/Error: require "google/apis/analyticsreporting_v4"
```