Use BigQuery for User Feedback analytics
Instead of fetching data from UA we now use the BigQuery API which
allows us to query the table for the specific metrics we need. It
already returns the information we need in the correct format and
reduces the amount of work required post query.

The test checking if yields successive report data if some actions
are missing defaulting to 0 was removed as this is no longer
required at this stage. This will be enforced by the DB.

https://trello.com/c/6OkkMnm0/3411-5-migrate-ua-metrics-for-user-feedback-to-bigquery-in-content-data-api

Co-authored-by:

- Murilo Dal Ri <murilo.dalri@digital.cabinet-office.gov.uk>
- Aga Dufrat <aga.dufrat@digital.cabinet-office.gov.uk>