Update Internal Search service to pull metrics from GA4
Instead of fetching data from UA we now use the BigQuery API which
allows us to query the table for the specific metrics we need from
GA4. It already returns the information we need in the correct format
and reduces the amount of work required post query.

`ga_response_with_govuk_prefix` has not been used in any of the tests and
looks like it has been missed previously. We should expect paths that
contain the gov.uk prefix to be transformed into it's base path. If an
an event already exists with the same path in the Events::GA table, it
should update the entry with the aggregative total for the metrics which is
then reflected in the Facts::Metrics table.