Fix flakey test
The `./spec/integration/streams/all_schemas_spec.rb` tests are flakey as
the payload sent to the message queue does not always contain a base
path.

This means the test passes on occasion when it shouldn't,  because the
application has been designed to silently ignore invalid payloads.

Explicitly adding a base path for these tests, to remove the flakiness.