Add `link_collection` schema to `NoContent` parser
A schema for `link_collection` was added in
https://github.com/alphagov/publishing-api/commit/e3cbefbb6a5adf60be555912f67431f207dad112
but a parser was not defined in this application, resulting in
subsequent CI builds failing.

Adding this to the `NoContent` parser as this not a page containing any
content, but rather something that is included in other pages to show a
list of links.