Add a test for presence of parser for every schema
It is possible to add a schema to `publishing-api` which has an optional
base path.

If a docment does not have a base path, this application silently
ignores the message and continues processing further jobs.

This leads to the other test here being flakey, since sometimes the
message will be silently dropped (because the random content generator
has created content without a base path) even though the test should be
failing as there is no relevant parser defined.

Therefore adding an additional test that ensures a parser is explicitly
defined for each schema type.