remove `content_block_email_address` type under NoContent
This type is being removed so we don't need to
account for it.

We are having to temporarily skip the schema name
in the integration test because the CI calls out
to the Publishing API to get the latest schema
names, and we can't delete the schema on the
Publishing API without first removing it here.